+++ core4/api/v1/request/standard/static.py
