+++ tests/api/test_info.py
@@ -2,7 +2,7 @@ import pytest
 
 from core4.api.v1.application import CoreApiContainer
 from core4.api.v1.request.main import CoreRequestHandler
-from core4.api.v1.request.static import CoreStaticFileHandler
+from core4.api.v1.request.standard.static import CoreStaticFileHandler
 from core4.api.v1.server import CoreApiServer
 from tests.api.test_test import setup, mongodb, run
 import core4
